Fix r2 build errors (#610)
* Fix r2 build errors

* Update

* Re-enable ROS bridge tests

* Update runs-on tags

* Why won't ROS run?

Because its dumb (though its not ROS's fault).

* Maybe now?

* Update main.yml

* getting ROS tests working

* Update main.yml

* Getting closer

* Enable -fPIC

* Weird recursive symlinks...

* Enable position-independent code

* more -fPIC (via cmake)

---------

+++ tests/runtime/device_runtime_tests.cpp
@@ -210,8 +210,8 @@ std::string printfToString(const char* format, ...)
 
 bool DeviceRuntimeTests::CheckGpsDuplicate(const char* description, int &count, uint32_t towMs, uint32_t gpsWeek, const uint8_t* msg, int msgSize, msg_history_t &hist)
 {
-    int toyMs = towMs + gpsWeek * C_MILLISECONDS_PER_WEEK;
-    int histToyMs = hist.gpsTowMs + hist.gpsWeek * C_MILLISECONDS_PER_WEEK;
+    int64_t toyMs = towMs + gpsWeek * C_MILLISECONDS_PER_WEEK;
+    int64_t histToyMs = hist.gpsTowMs + hist.gpsWeek * C_MILLISECONDS_PER_WEEK;
 
     if (toyMs == histToyMs)
     {   // Duplicate time
@@ -229,8 +229,8 @@ bool DeviceRuntimeTests::CheckGpsDuplicate(const char* description, int &count, 
 
 bool DeviceRuntimeTests::CheckGpsTimeReverse(const char* description, int &count, uint32_t towMs, uint32_t gpsWeek, const uint8_t* msg, int msgSize, msg_history_t &hist)
 {
-    int toyMs = towMs + gpsWeek * C_MILLISECONDS_PER_WEEK;
-    int histToyMs = hist.gpsTowMs + hist.gpsWeek * C_MILLISECONDS_PER_WEEK;
+    int64_t toyMs = towMs + gpsWeek * C_MILLISECONDS_PER_WEEK;
+    int64_t histToyMs = hist.gpsTowMs + hist.gpsWeek * C_MILLISECONDS_PER_WEEK;
 
     if (toyMs < histToyMs)
     {   // Reversed time
